# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Document status
 */
export type RegulatoryDocumentStatus =
  (typeof RegulatoryDocumentStatus)[keyof typeof RegulatoryDocumentStatus];

export const RegulatoryDocumentStatus = {
  Current: "Current",
  Expiring_Soon: "Expiring Soon",
  Expired: "Expired",
  Unknown: "Unknown",
} as const;

export interface RegulatoryDocument {
  /** Notion page ID */
  id: string;
  /** Document name */
  name: string;
  /**
   * Expiration/renewal date (ISO 8601)
   * @nullable
   */
  expirationDate?: string | null;
  /** Document status */
  status: RegulatoryDocumentStatus;
  /**
   * Google Drive file URL
   * @nullable
   */
  fileLink?: string | null;
  /**
   * Days until expiration (negative = already expired)
   * @nullable
   */
  daysUntilExpiration?: number | null;
  /** Whether a Google Calendar reminder event has been created */
  calendarEventCreated: boolean;
}

export interface RegulatorySummary {
  total: number;
  current: number;
  expiringSoon: number;
  expired: number;
  notionsConnected: boolean;
}

export interface CalendarSyncResult {
  eventsCreated: number;
  eventsSkipped: number;
  errors: string[];
}

export interface AppSettings {
  notionRegulatoryDbId: string;
  googleCalendarId: string;
  notionAeLogDbId: string;
  notionDeviationLogDbId: string;
  googleSheetsId: string;
  googleSheetTab: string;
  googleSheetHeaderRow: number;
  googleDocsTemplateId: string;
  sponsorCallEventId: string;
  piEmail: string;
  sponsorEmail: string;
  nagIntervalHours: number;
}

export interface AppSettingsInput {
  notionRegulatoryDbId?: string;
  googleCalendarId?: string;
  notionAeLogDbId?: string;
  notionDeviationLogDbId?: string;
  googleSheetsId?: string;
  googleSheetTab?: string;
  googleSheetHeaderRow?: number;
  googleDocsTemplateId?: string;
  sponsorCallEventId?: string;
  piEmail?: string;
  sponsorEmail?: string;
  nagIntervalHours?: number;
}

export type SponsorReportStatus =
  (typeof SponsorReportStatus)[keyof typeof SponsorReportStatus];

export const SponsorReportStatus = {
  Draft: "Draft",
  PI_Review: "PI Review",
  Approved: "Approved",
  Sent: "Sent",
  Discarded: "Discarded",
} as const;

export interface SponsorReport {
  /** UUID */
  id: string;
  /** Google Docs URL */
  docUrl: string;
  /** Google Drive file ID */
  docId: string;
  status: SponsorReportStatus;
  /** ISO 8601 timestamp */
  generatedAt: string;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  sentToPiAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  lastNagAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  approvedAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  finalizedAt?: string | null;
  /** Any {{...}} placeholders not replaced during generation */
  unreplacedPlaceholders: string[];
}

export interface GenerateReportInput {
  /** Set true to proceed despite stale enrollment data */
  acknowledgeStale?: boolean;
}

export interface GenerateReportResponse {
  report?: SponsorReport;
  /**
   * Warning if sheet not updated in last 24h
   * @nullable
   */
  stalenessWarning?: string | null;
  /** If true, re-submit with acknowledgeStale=true to proceed */
  requiresStaleAcknowledge: boolean;
}

export interface NagCheckResponse {
  nagsSent: number;
  errors: string[];
}

export interface ErrorResponse {
  error: string;
}