# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: regulatory
    description: Regulatory document management
  - name: settings
    description: Application settings
  - name: reports
    description: Sponsor report co-pilot
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /regulatory/documents:
    get:
      operationId: listRegulatoryDocuments
      tags: [regulatory]
      summary: List regulatory documents
      description: Returns all regulatory documents from Notion sorted by expiration date
      responses:
        "200":
          description: List of regulatory documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegulatoryDocument"
        "503":
          description: Notion not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /regulatory/summary:
    get:
      operationId: getRegulatorysummary
      tags: [regulatory]
      summary: Regulatory document summary
      description: Returns counts by status for dashboard display
      responses:
        "200":
          description: Summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulatorySummary"

  /regulatory/sync-calendar:
    post:
      operationId: syncRegulatoryCalendar
      tags: [regulatory]
      summary: Sync calendar reminders
      description: Creates Google Calendar reminder events for documents expiring within 30 days
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSyncResult"
        "503":
          description: Google Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get application settings
      description: Returns current application configuration
      responses:
        "200":
          description: Settings object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update application settings
      description: Saves application configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List sponsor reports
      description: Returns all sponsor report records sorted by generation date (newest first)
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SponsorReport"

  /reports/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate sponsor report
      description: Fetches data from Sheets/Notion, copies the template Doc, fills placeholders, grants PI access. Returns staleness warning if sheet >24h old.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportInput"
      responses:
        "200":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateReportResponse"
        "409":
          description: Concurrent report guard — active report in PI Review or Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Required integration not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{reportId}/send-to-pi:
    post:
      operationId: sendReportToPi
      tags: [reports]
      summary: Send report to PI for review
      description: Sends Gmail to PI with doc link and advances status to PI Review
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email sent, status advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReport"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{reportId}/mark-approved:
    post:
      operationId: markReportApproved
      tags: [reports]
      summary: Mark report as PI approved
      description: Manually advances report status from PI Review to Approved
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status advanced to Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReport"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{reportId}/mark-final:
    post:
      operationId: markReportFinal
      tags: [reports]
      summary: Mark report final and update sponsor call
      description: Appends report link to calendar event description and advances status to Sent
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status advanced to Sent, calendar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReport"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{reportId}/discard:
    post:
      operationId: discardReport
      tags: [reports]
      summary: Discard draft report
      description: Deletes the orphaned Google Doc from Drive and marks report as Discarded
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report discarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReport"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /internal/nag-check:
    post:
      operationId: nagCheck
      tags: [reports]
      summary: PI nag engine check
      description: Called by Replit Scheduled Deployment every 4 hours. Sends follow-up Gmail to PI for any reports still in PI Review past the nag interval.
      responses:
        "200":
          description: Nag check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NagCheckResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegulatoryDocument:
      type: object
      properties:
        id:
          type: string
          description: Notion page ID
        name:
          type: string
          description: Document name
        expirationDate:
          type: ["string", "null"]
          description: Expiration/renewal date (ISO 8601)
        status:
          type: string
          enum: [Current, "Expiring Soon", Expired, Unknown]
          description: Document status
        fileLink:
          type: ["string", "null"]
          description: Google Drive file URL
        daysUntilExpiration:
          type: ["integer", "null"]
          description: Days until expiration (negative = already expired)
        calendarEventCreated:
          type: boolean
          description: Whether a Google Calendar reminder event has been created
      required:
        - id
        - name
        - status
        - calendarEventCreated

    RegulatorySummary:
      type: object
      properties:
        total:
          type: integer
        current:
          type: integer
        expiringSoon:
          type: integer
        expired:
          type: integer
        notionsConnected:
          type: boolean
      required:
        - total
        - current
        - expiringSoon
        - expired
        - notionsConnected

    CalendarSyncResult:
      type: object
      properties:
        eventsCreated:
          type: integer
        eventsSkipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - eventsCreated
        - eventsSkipped
        - errors

    AppSettings:
      type: object
      properties:
        notionRegulatoryDbId:
          type: string
        googleCalendarId:
          type: string
        notionAeLogDbId:
          type: string
        notionDeviationLogDbId:
          type: string
        googleSheetsId:
          type: string
        googleSheetTab:
          type: string
        googleSheetHeaderRow:
          type: integer
        googleDocsTemplateId:
          type: string
        sponsorCallEventId:
          type: string
        piEmail:
          type: string
        sponsorEmail:
          type: string
        nagIntervalHours:
          type: integer
      required:
        - notionRegulatoryDbId
        - googleCalendarId
        - notionAeLogDbId
        - notionDeviationLogDbId
        - googleSheetsId
        - googleSheetTab
        - googleSheetHeaderRow
        - googleDocsTemplateId
        - sponsorCallEventId
        - piEmail
        - sponsorEmail
        - nagIntervalHours

    AppSettingsInput:
      type: object
      properties:
        notionRegulatoryDbId:
          type: string
        googleCalendarId:
          type: string
        notionAeLogDbId:
          type: string
        notionDeviationLogDbId:
          type: string
        googleSheetsId:
          type: string
        googleSheetTab:
          type: string
        googleSheetHeaderRow:
          type: integer
        googleDocsTemplateId:
          type: string
        sponsorCallEventId:
          type: string
        piEmail:
          type: string
        sponsorEmail:
          type: string
        nagIntervalHours:
          type: integer

    SponsorReport:
      type: object
      properties:
        id:
          type: string
          description: UUID
        docUrl:
          type: string
          description: Google Docs URL
        docId:
          type: string
          description: Google Drive file ID
        status:
          type: string
          enum: [Draft, "PI Review", Approved, Sent, Discarded]
        generatedAt:
          type: string
          description: ISO 8601 timestamp
        sentToPiAt:
          type: ["string", "null"]
          description: ISO 8601 timestamp
        lastNagAt:
          type: ["string", "null"]
          description: ISO 8601 timestamp
        approvedAt:
          type: ["string", "null"]
          description: ISO 8601 timestamp
        finalizedAt:
          type: ["string", "null"]
          description: ISO 8601 timestamp
        unreplacedPlaceholders:
          type: array
          items:
            type: string
          description: Any {{...}} placeholders not replaced during generation
      required:
        - id
        - docUrl
        - docId
        - status
        - generatedAt
        - unreplacedPlaceholders

    GenerateReportInput:
      type: object
      properties:
        acknowledgeStale:
          type: boolean
          description: Set true to proceed despite stale enrollment data
      required: []

    GenerateReportResponse:
      type: object
      properties:
        report:
          $ref: "#/components/schemas/SponsorReport"
        stalenessWarning:
          type: ["string", "null"]
          description: Warning if sheet not updated in last 24h
        requiresStaleAcknowledge:
          type: boolean
          description: If true, re-submit with acknowledgeStale=true to proceed
      required:
        - requiresStaleAcknowledge

    NagCheckResponse:
      type: object
      properties:
        nagsSent:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - nagsSent
        - errors

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error