# lib/api-zod/src/generated/types/generateReportInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateReportInput {
  /** Set true to proceed despite stale enrollment data */
  acknowledgeStale?: boolean;
}