# lib/api-zod/src/generated/types/generateReportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorReport } from "./sponsorReport";

export interface GenerateReportResponse {
  report?: SponsorReport;
  /**
   * Warning if sheet not updated in last 24h
   * @nullable
   */
  stalenessWarning?: string | null;
  /** If true, re-submit with acknowledgeStale=true to proceed */
  requiresStaleAcknowledge: boolean;
}