# lib/api-zod/src/generated/types/sponsorReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorReportStatus } from "./sponsorReportStatus";

export interface SponsorReport {
  /** UUID */
  id: string;
  /** Google Docs URL */
  docUrl: string;
  /** Google Drive file ID */
  docId: string;
  status: SponsorReportStatus;
  /** ISO 8601 timestamp */
  generatedAt: string;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  sentToPiAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  lastNagAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  approvedAt?: string | null;
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  finalizedAt?: string | null;
  /** Any {{...}} placeholders not replaced during generation */
  unreplacedPlaceholders: string[];
}